Removed pending handler from Component API
